Use current culture for CSV
Add -UseCulture to the Export-CSV command.
Because the generated file will most likely be opened using Excel,
which expects the field delimiters according on the current culture.

e.g. a German Excel expects ";" as field delimiter.

Also, I added a .gitignore file which will ignore .csv files in
the working tree, as a run of the script saves the result as
a local file, too.